backend/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
